+++ src/frontend/User.java
@@ -39,9 +39,9 @@ public class User {
     }
 
     public double getCantEnPeso() { //devuelve la suma de cantidades en kg
-        int suma = 0;
+        double suma = 0;
         for (int i=0; i < coleccionDeAcopio.size(); i++) {
-            suma += (coleccionDeAcopio.elementAt(i).getSecond()* coleccionDeAcopio.elementAt(i).getFirst().getPeso());
+        	suma += (coleccionDeAcopio.elementAt(i).getSecond()* (double)coleccionDeAcopio.elementAt(i).getFirst().getPeso());
         }
         return suma;
     }
